Cue becomes attribute of CueRender
